expiry selection in shareSettingsPopover.tsx
