+++ app/components/Sharing/Collection/PublicAccess.tsx
@@ -5,8 +5,7 @@ import { CopyIcon, GlobeIcon } from "outline-icons";
 import * as React from "react";
 import { useTranslation } from "react-i18next";
 import { toast } from "sonner";
-import styled, { useTheme } from "styled-components";
-import { differenceInCalendarDays, endOfDay } from "date-fns";
+import { useTheme } from "styled-components";
 import { errToString } from "@shared/utils/error";
 import Squircle from "@shared/components/Squircle";
 import { UrlHelper } from "@shared/utils/UrlHelper";
@@ -15,7 +14,6 @@ import type Share from "~/models/Share";
 import { AvatarSize } from "~/components/Avatar";
 import CopyToClipboard from "~/components/CopyToClipboard";
 import Flex from "~/components/Flex";
-import { InputSelect, type Option } from "~/components/InputSelect";
 import NudeButton from "~/components/NudeButton";
 import { ResizingHeightContainer } from "~/components/ResizingHeightContainer";
 import Switch from "~/components/Switch";
@@ -24,17 +22,9 @@ import Tooltip from "~/components/Tooltip";
 import env from "~/env";
 import usePolicy from "~/hooks/usePolicy";
 import useStores from "~/hooks/useStores";
-import useUserLocale from "~/hooks/useUserLocale";
-import { dateToExpiry } from "~/utils/date";
 import { ListItem } from "../components/ListItem";
 import ShareSettingsPopover from "../components/ShareSettingsPopover";
 import { DomainPrefix, ShareLinkInput, StyledInfoIcon } from "../components";
-import ExpiryDatePicker from "~/scenes/ApiKeyNew/components/ExpiryDatePicker";
-import {
-  ExpiryType,
-  ExpiryValues,
-  calculateExpiryDate,
-} from "~/scenes/ApiKeyNew/utils";
 
 type Props = {
   /** The collection to share. */
@@ -57,52 +47,10 @@ function InnerPublicAccess(
   const collectionAbilities = usePolicy(collection);
   const canPublish = share ? can.update : collectionAbilities.share;
   const [creating, setCreating] = React.useState(false);
-  const userLocale = useUserLocale();
-  const currentDate = React.useRef<Date>(new Date());
-  const [expiryType, setExpiryType] = React.useState<ExpiryType>(
-    ExpiryType.NoExpiration
-  );
-  const [expiresAt, setExpiresAt] = React.useState<Date | undefined>(
-    new Date()
-  );
-
-  const expiryOptions = React.useMemo<Option[]>(
-    () =>
-      [...ExpiryValues.entries()].map(([expType, { label }]) => ({
-        type: "item",
-        label,
-        value: expType,
-      })),
-    []
-  );
 
   React.useEffect(() => {
     setUrlId(share?.urlId);
   }, [share?.urlId]);
-
-  React.useEffect(() => {
-    if (!share?.expiresAt) {
-      setExpiryType(ExpiryType.NoExpiration);
-      setExpiresAt(undefined);
-      return;
-    }
-
-    const parsedExpiresAt = new Date(share.expiresAt);
-    const daysUntilExpiry = differenceInCalendarDays(
-      parsedExpiresAt,
-      currentDate.current
-    );
-
-    const matchedType = [
-      ExpiryType.Week,
-      ExpiryType.Month,
-      ExpiryType.TwoMonths,
-      ExpiryType.ThreeMonths,
-    ].find((type) => ExpiryValues.get(type)?.value === daysUntilExpiry);
-
-    setExpiresAt(parsedExpiresAt);
-    setExpiryType(matchedType ?? ExpiryType.Custom);
-  }, []);
 
   const handlePublishedChange = React.useCallback(
     async (checked: boolean) => {
@@ -113,15 +61,11 @@ function InnerPublicAccess(
             type: "collection",
             collectionId: collection.id,
             published: true,
-            expiresAt: expiresAt?.toISOString(),
           });
           copy(newShare.url);
           toast.success(t("Public link copied to clipboard"));
         } else if (share) {
-          await share.save({
-            published: checked,
-            expiresAt: expiresAt?.toISOString() ?? null,
-          });
+          await share.save({ published: checked });
           if (checked) {
             copy(share.url);
             toast.success(t("Public link copied to clipboard"));
@@ -133,46 +77,7 @@ function InnerPublicAccess(
         setCreating(false);
       }
     },
-    [t, share, shares, collection, expiresAt]
-  );
-
-  const updateShareExpiry = React.useCallback(
-    async (date?: Date) => {
-      if (!share?.published) {
-        return;
-      }
-
-      try {
-        await share.save({ expiresAt: date?.toISOString() ?? null });
-      } catch (err) {
-        toast.error(errToString(err));
-      }
-    },
-    [share]
-  );
-
-  const handleExpiryTypeChange = React.useCallback(
-    async (value: string) => {
-      const nextExpiryType = value as ExpiryType;
-      const nextExpiresAt = calculateExpiryDate(
-        currentDate.current,
-        nextExpiryType
-      );
-
-      setExpiryType(nextExpiryType);
-      setExpiresAt(nextExpiresAt);
-      await updateShareExpiry(nextExpiresAt);
-    },
-    [updateShareExpiry]
-  );
-
-  const handleSelectCustomDate = React.useCallback(
-    async (date: Date) => {
-      const nextExpiresAt = endOfDay(date);
-      setExpiresAt(nextExpiresAt);
-      await updateShareExpiry(nextExpiresAt);
-    },
-    [updateShareExpiry]
+    [t, share, shares, collection]
   );
 
   const handleUrlChange = React.useMemo(
@@ -264,26 +169,6 @@ function InnerPublicAccess(
               </ShareLinkInput>
               <ShareSettingsPopover share={share} />
             </Flex>
-            <Flex align="center" gap={8}>
-              <StyledExpirySelect
-                options={expiryOptions}
-                value={expiryType}
-                onChange={handleExpiryTypeChange}
-                label={t("Link Expiration")}
-              />
-              {expiryType === ExpiryType.Custom ? (
-                <ExpiryDatePicker
-                  selectedDate={expiresAt}
-                  onSelect={handleSelectCustomDate}
-                />
-              ) : (
-                <Text type="secondary" size="small">
-                  {expiresAt
-                    ? `${dateToExpiry(expiresAt.toISOString(), t, userLocale)}.`
-                    : `${t("Never expires")}.`}
-                </Text>
-              )}
-            </Flex>
             <Flex align="flex-start" gap={4}>
               <StyledInfoIcon color={theme.textTertiary} />
               <Text type="tertiary" size="xsmall">
@@ -299,10 +184,5 @@ function InnerPublicAccess(
     </div>
   );
 }
-
-const StyledExpirySelect = styled(InputSelect)`
-  width: 150px !important;
-  margin-bottom: 0;
-`;
 
 export const PublicAccess = observer(React.forwardRef(InnerPublicAccess));
