+++ app/components/Sharing/Document/PublicAccess.tsx
@@ -5,21 +5,17 @@ import { CopyIcon, GlobeIcon } from "outline-icons";
 import * as React from "react";
 import { Trans, useTranslation } from "react-i18next";
 import { toast } from "sonner";
-import styled, { useTheme } from "styled-components";
-import { differenceInCalendarDays, endOfDay } from "date-fns";
+import { useTheme } from "styled-components";
 import { errToString } from "@shared/utils/error";
 import Flex from "@shared/components/Flex";
 import Squircle from "@shared/components/Squircle";
 import { UrlHelper } from "@shared/utils/UrlHelper";
 import type Document from "~/models/Document";
 import type Share from "~/models/Share";
-import { InputSelect, type Option } from "~/components/InputSelect";
 import Switch from "~/components/Switch";
 import env from "~/env";
 import usePolicy from "~/hooks/usePolicy";
 import useStores from "~/hooks/useStores";
-import useUserLocale from "~/hooks/useUserLocale";
-import { dateToExpiry } from "~/utils/date";
 import { AvatarSize } from "../../Avatar";
 import CopyToClipboard from "../../CopyToClipboard";
 import NudeButton from "../../NudeButton";
@@ -34,13 +30,6 @@ import {
   StyledInfoIcon,
   UnderlinedLink,
 } from "../components";
-import ExpiryDatePicker from "~/scenes/ApiKeyNew/components/ExpiryDatePicker";
-import {
-  ExpiryType,
-  ExpiryValues,
-  calculateExpiryDate,
-} from "~/scenes/ApiKeyNew/utils";
-import Container from "~/scenes/Document/components/Container";
 
 type Props = {
   /** The document to share. */
@@ -68,52 +57,10 @@ function PublicAccess(
   const documentAbilities = usePolicy(document);
   const canPublish = share ? can.update : documentAbilities.share;
   const [creating, setCreating] = React.useState(false);
-  const userLocale = useUserLocale();
-  const currentDate = React.useRef<Date>(new Date());
-  const [expiryType, setExpiryType] = React.useState<ExpiryType>(
-    ExpiryType.NoExpiration
-  );
-  const [expiresAt, setExpiresAt] = React.useState<Date | undefined>(
-    new Date()
-  );
-
-  const expiryOptions = React.useMemo<Option[]>(
-    () =>
-      [...ExpiryValues.entries()].map(([expType, { label }]) => ({
-        type: "item",
-        label,
-        value: expType,
-      })),
-    []
-  );
 
   React.useEffect(() => {
     setUrlId(share?.urlId);
   }, [share?.urlId]);
-
-  React.useEffect(() => {
-    if (!share?.expiresAt) {
-      setExpiryType(ExpiryType.NoExpiration);
-      setExpiresAt(undefined);
-      return;
-    }
-
-    const parsedExpiresAt = new Date(share.expiresAt);
-    const daysUntilExpiry = differenceInCalendarDays(
-      parsedExpiresAt,
-      currentDate.current
-    );
-
-    const matchedType = [
-      ExpiryType.Week,
-      ExpiryType.Month,
-      ExpiryType.TwoMonths,
-      ExpiryType.ThreeMonths,
-    ].find((type) => ExpiryValues.get(type)?.value === daysUntilExpiry);
-
-    setExpiresAt(parsedExpiresAt);
-    setExpiryType(matchedType ?? ExpiryType.Custom);
-  }, []);
 
   const handlePublishedChange = React.useCallback(
     async (checked: boolean) => {
@@ -124,15 +71,11 @@ function PublicAccess(
             type: "document",
             documentId: document.id,
             published: true,
-            expiresAt: expiresAt?.toISOString(),
           });
           copy(newShare.url);
           toast.success(t("Public link copied to clipboard"));
         } else if (share) {
-          await share.save({
-            published: checked,
-            expiresAt: expiresAt?.toISOString() ?? null,
-          });
+          await share.save({ published: checked });
           if (checked) {
             copy(share.url);
             toast.success(t("Public link copied to clipboard"));
@@ -144,46 +87,7 @@ function PublicAccess(
         setCreating(false);
       }
     },
-    [t, share, shares, document, expiresAt]
-  );
-
-  const updateShareExpiry = React.useCallback(
-    async (date?: Date) => {
-      if (!share?.published) {
-        return;
-      }
-
-      try {
-        await share.save({ expiresAt: date?.toISOString() ?? null });
-      } catch (err) {
-        toast.error(errToString(err));
-      }
-    },
-    [share]
-  );
-
-  const handleExpiryTypeChange = React.useCallback(
-    async (value: string) => {
-      const nextExpiryType = value as ExpiryType;
-      const nextExpiresAt = calculateExpiryDate(
-        currentDate.current,
-        nextExpiryType
-      );
-
-      setExpiryType(nextExpiryType);
-      setExpiresAt(nextExpiresAt);
-      await updateShareExpiry(nextExpiresAt);
-    },
-    [updateShareExpiry]
-  );
-
-  const handleSelectCustomDate = React.useCallback(
-    async (date: Date) => {
-      const nextExpiresAt = endOfDay(date);
-      setExpiresAt(nextExpiresAt);
-      await updateShareExpiry(nextExpiresAt);
-    },
-    [updateShareExpiry]
+    [t, share, shares, document]
   );
 
   const handleUrlChange = React.useMemo(
@@ -294,48 +198,24 @@ function PublicAccess(
             {copyButton}
           </ShareLinkInput>
         ) : share?.published ? (
-          <>
-            <Flex align="center" gap={2}>
-              <ShareLinkInput
-                type="text"
-                ref={inputRef}
-                placeholder={share?.id}
-                onChange={handleUrlChange}
-                error={validationError}
-                defaultValue={urlId}
-                prefix={
-                  <DomainPrefix onClick={() => inputRef.current?.focus()}>
-                    {env.URL.replace(/https?:\/\//, "") + "/s/"}
-                  </DomainPrefix>
-                }
-              >
-                {copyButton}
-              </ShareLinkInput>
-              <ShareSettingsPopover share={share} />
-            </Flex>
-            <Flex align="center" gap={8}>
-              <StyledExpirySelect
-                options={expiryOptions}
-                value={expiryType}
-                onChange={handleExpiryTypeChange}
-                label={t("Link Expiration")}
-              />
-              {expiryType === ExpiryType.Custom ? (
-                <Container style={{ marginTop: 16 }}>
-                  <ExpiryDatePicker
-                    selectedDate={expiresAt}
-                    onSelect={handleSelectCustomDate}
-                  />
-                </Container>
-              ) : (
-                <Text type="secondary" size="small" style={{ marginTop: 25 }}>
-                  {expiresAt
-                    ? `${dateToExpiry(expiresAt.toISOString(), t, userLocale)}.`
-                    : `${t("Never expires")}.`}
-                </Text>
-              )}
-            </Flex>
-          </>
+          <Flex align="center" gap={2}>
+            <ShareLinkInput
+              type="text"
+              ref={inputRef}
+              placeholder={share?.id}
+              onChange={handleUrlChange}
+              error={validationError}
+              defaultValue={urlId}
+              prefix={
+                <DomainPrefix onClick={() => inputRef.current?.focus()}>
+                  {env.URL.replace(/https?:\/\//, "") + "/s/"}
+                </DomainPrefix>
+              }
+            >
+              {copyButton}
+            </ShareLinkInput>
+            <ShareSettingsPopover share={share} />
+          </Flex>
         ) : null}
 
         {share?.published && !share.includeChildDocuments ? (
@@ -353,11 +233,5 @@ function PublicAccess(
     </div>
   );
 }
-
-const StyledExpirySelect = styled(InputSelect)`
-  width: 150px !important;
-  margin-bottom: 0;
-  margin-top: 0;
-`;
 
 export default observer(React.forwardRef(PublicAccess));
