+++ shared/i18n/locales/en_US/translation.json
@@ -445,8 +445,6 @@
   "Publish to web": "Publish to web",
   "Allow anyone with the link to access": "Allow anyone with the link to access",
   "Publish to internet": "Publish to internet",
-  "Link Expiration": "Link Expiration",
-  "Never expires": "Never expires",
   "All documents in this collection will be shared on the web, including any new documents added later": "All documents in this collection will be shared on the web, including any new documents added later",
   "Invite": "Invite",
   "{{ userName }} was added to the collection": "{{ userName }} was added to the collection",
@@ -478,6 +476,9 @@
   "Disable this setting to discourage search engines from indexing the page": "Disable this setting to discourage search engines from indexing the page",
   "Email subscriptions": "Email subscriptions",
   "Allow viewers to subscribe and receive email notifications when documents are updated": "Allow viewers to subscribe and receive email notifications when documents are updated",
+  "Link Expiration": "Link Expiration",
+  "Set an expiration date for the link": "Set an expiration date for the link",
+  "Never expires": "Never expires",
   "Something went wrong": "Something went wrong",
   "Check your email to confirm your subscription": "Check your email to confirm your subscription",
   "Get notified when this document is updated": "Get notified when this document is updated",
